requirements.txt
gspread
oauth2client
pandas
pyyaml
tqdm
google-cloud-bigquery
db-dtypes
ydata-profiling
deep-translator
langdetect
transformers
torch
sentence-transformers==2.2.2
pymongo
openai
openpyxl
datasets